test now runs a systems check organized basecalls gyro center gear perfectted renamed and moved some suty stuff around

+++ Trephine/UtilityFunctions.cs
@@ -22,7 +22,7 @@
         /// <param name="original"></param>
         /// <param name="turn"></param>
         /// <returns></returns>
-        public static int GetDestination(int original, int turn)
+        public static double GetDestination(double original, double turn)
         {
             var val = original + turn;
